+++ bot/Dockerfile
@@ -1,10 +1,12 @@
-FROM python:3.7-slim
+FROM python:3.9-alpine
+
+RUN apk update && apk add python3-dev gcc libc-dev mysql-client build-base
 
 COPY . /bot
 
 WORKDIR /bot
 
-RUN pip3.7 install -r requirements.txt
+RUN pip3.9 install -r requirements.txt
 
-ENTRYPOINT ["python3.7"]
+ENTRYPOINT ["python3.9"]
 CMD ["main.py"]
