+++ bot/core/cogs/__init__.py
@@ -1,0 +1,1 @@
+from .webhooks import Webhooks
