+++ bot/core/tools.py
